fix for #26670 ensure Windows paths use native separators for zip files
The setup_project command assumes that the path to an existing config on disk is a native path (using the proper slash direction). When selecting a zip file, QT is always returning paths with forward slashes. This fix ensures that Windows paths have them converted to backslashes.

Tweaked the layout a little bit to tighten up the button spacing.